SLING-7788: restructured MockServiceRegistration
To add the services in the osgi metadata to the properties of the
service registration as well, the osgi metadata are nowread before
the properties map is assembled. This will add any passed service
classes and those in the metadata to the objectClass service property
 and so will make them available to ServiceListeners which the
ServiceTracker is using to track services.